Extract interface for Content
